+++ R/percent_missing.R
@@ -20,6 +20,13 @@ percent_missing <- function(df, grouping_cols = NULL,exclude_cols = NULL){
 }
 
 #' @export
+
+percent_missing.default <- function(df, grouping_cols = NULL,exclude_cols = NULL){
+  stop(paste0("percent_missing not available for objects of class ",
+              class(df)[1], ". Try 'percent_na(x)' instead."))
+}
+
+#' @export
 percent_missing.data.frame <- function(df,  grouping_cols = NULL,
                                        exclude_cols = NULL){
 
@@ -37,7 +44,7 @@ if(! is.null(exclude_cols)){
 check_column_existence(df, exclude_cols, "to exclude")
 
 
-  df <- df %>% dplyr::select(-exclude_cols)
+  df <- df %>% dplyr::select(-dplyr::all_of(exclude_cols))
 
 
 }
@@ -46,12 +53,21 @@ check_column_existence(df, exclude_cols, "to exclude")
                                  ~ get_na_means(.))) %>% dplyr::ungroup()
 
 }
+#' percent missing but for vectors. 
+#' @inheritParams get_na_means
+#' @examples 
+#' percent_na(airquality$Ozone)
+#' @export
 
-percent_missing.numeric <- get_na_means.numeric
-
-percent_missing.factor <- get_na_means.character
-
-percent_missing.character <- get_na_means.character
+percent_na <- function(x) UseMethod("percent_na")
+#' @export 
+percent_na.numeric <- function(x) mean(is.na(x)) * 100
+#' @export
+percent_na.factor <- percent_na.numeric 
+#' @export
+percent_na.character <- percent_na.numeric 
+#' @export
+percent_na.POSIXct <- percent_na.numeric