+++ tests/testthat/test_helpers.R
@@ -1,5 +1,8 @@
-testthat::test_that(desc="Test helpers",
+test_that(desc = "Test helpers",
                     code = {
-expect_true(get_na_means.data.frame(airquality)[[1]]==percent_missing.data.frame(airquality)[[1]])
-
+skip_on_oldrel()
+expect_true(
+get_na_means.data.frame(airquality)[[1]] == percent_missing.data.frame(airquality)[[1]]
+                      )
+                      
                     })